requirements.txt
# pip install -r requirements.txt
# tested on Ubuntu 24.04.1 LTS and Python 3.12
#
# to run MOP+MiHo+NCC demo script, aka miho_ncc_demo.py
numpy==2.0.2
scipy==1.14.1
kornia==0.7.4
kornia_rs==0.1.7
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.0.0
matplotlib==3.9.3
PyQt6==6.7.1
# 
# to run the benchmark script too
tensorflow-cpu==2.18.0
# pydegensac cannot be installed, removed until a fix is found
# pydegensac==0.1.2
gdown==5.2.0
rich==13.9.4
sympy==1.13.1
munkres==1.1.4
einops==0.8.0
e2cnn==0.2.3
parse==1.20.2
pycolmap==3.11.0
xformers==0.0.28.post3
poselib==2.0.4
tqdm>=4.67.1
tzdata==2024.2
zipp>=3.21.0
lightglue @ git+https://github.com/cvg/LightGlue
romatch
# progressive-x can be installed from git: https://github.com/danini/progressive-x/tree/master
# key2subpx git: https://github.com/KimSinjeong/keypt2subpx
# pymulticonsensus cannot be installed, removed until a fix is found
# pymulticonsensus @ git+https://github.com/danini/clustering-in-consensus-space
#
# to activate DIM modules
# deep-image-matching @ git+https://github.com/3DOM-FBK/deep-image-matching
